+++ packages/runtime/src/ctx.ts
@@ -4,13 +4,15 @@ import { AssetDef, makeAssetTransferTxnWithSuggestedParams } from "algosdk";
 import { Runtime } from ".";
 import { RUNTIME_ERRORS } from "./errors/errors-list";
 import { RuntimeError } from "./errors/runtime-errors";
+import { Err } from "./interpreter/opcode-list";
+import { ALGORAND_MIN_TX_FEE } from "./lib/constants";
 import { mockSuggestedParams } from "./mock/tx";
 import {
   AccountAddress, AccountStoreI,
   AppDeploymentFlags,
   ASADeploymentFlags, AssetHoldingM,
   Context, ExecutionMode,
-  SSCAttributesM, State, Txn, TxParams
+  SSCAttributesM, State, Txn
 } from "./types";
 
 const APPROVAL_PROGRAM = "approval-program";
@@ -174,7 +176,7 @@ export class Ctx implements Context {
    * @param address Account address to opt-into asset
    * @param flags Transaction Parameters
    */
-  optIntoASA (assetIndex: number, address: AccountAddress, flags: TxParams): void {
+  optIntoASA (assetIndex: number, address: AccountAddress, flags: types.TxParams): void {
     const assetDef = this.getAssetDef(assetIndex);
     makeAssetTransferTxnWithSuggestedParams(
       address, address, undefined, undefined, 0, undefined, assetIndex,
@@ -261,13 +263,45 @@ export class Ctx implements Context {
   }
 
   /**
+   * Verify Pooled Transaction Fees
+   * supports pooled fees where one transaction can pay the
+   * fees of other transactions within an atomic group.
+   * For atomic transactions, the protocol sums the number of
+   * transactions and calculates the total amount of required fees,
+   * then calculates the amount of fees submitted by all transactions.
+   * If the collected fees are greater than or equal to the required amount,
+   * the transaction fee requirement will be met.
+   * https://developer.algorand.org/articles/introducing-algorand-virtual-machine-avm-09-release/
+   */
+  verifyMinimumFees (): void {
+    let collected = 0;
+    for (const val of this.gtxs) {
+      if (val.fee === undefined) val.fee = 0;
+      collected += val.fee;
+    }
+
+    const required = this.gtxs.length * ALGORAND_MIN_TX_FEE;
+    if (collected < required) {
+      throw new RuntimeError(RUNTIME_ERRORS.TRANSACTION.FEES_NOT_ENOUGH, {
+        required: required,
+        collected: collected
+      });
+    }
+  }
+
+  /**
    * Deduct transaction fee from sender account.
    * @param sender Sender address
    * @param index Index of current tx being processed in tx group
    */
-  deductFee (sender: AccountAddress, index: number): void {
+  deductFee (sender: AccountAddress, index: number, params: types.TxParams): void {
+    let fee;
+    // If flatFee boolean is not set, change fee value
+    if (!params.flatFee && params.totalFee === undefined) {
+      fee = Math.max(ALGORAND_MIN_TX_FEE, this.gtxs[index].fee);
+    }
     const fromAccount = this.getAccount(sender);
-    const fee = BigInt(this.gtxs[index].fee);
+    fee = BigInt(this.gtxs[index].fee);
     fromAccount.amount -= fee; // remove tx fee from Sender's account
     this.assertAccBalAboveMin(fromAccount.address);
   }
@@ -450,9 +484,10 @@ export class Ctx implements Context {
    */
   /* eslint-disable sonarjs/cognitive-complexity */
   processTransactions (txnParams: types.ExecParams[]): void {
+    this.verifyMinimumFees();
     txnParams.forEach((txnParam, idx) => {
       const fromAccountAddr = webTx.getFromAddress(txnParam);
-      this.deductFee(fromAccountAddr, idx);
+      this.deductFee(fromAccountAddr, idx, txnParam.payFlags);
 
       if (txnParam.sign === types.SignType.LogicSignature) {
         this.tx = this.gtxs[idx]; // update current tx to index of stateless